Mollify GCC.

+++ c++/src/kj/common.h
@@ -429,8 +429,8 @@ static constexpr MinValue_ minValue = MinValue_();
 //
 // `char` is not supported, but `signed char` and `unsigned char` are.
 
-inline double inf() { return 1.0 / 0.0; }
-inline double nan() { return 0.0 / 0.0; }
+inline float inf() { return 1.0f / 0.0f; }
+inline float nan() { return 0.0f / 0.0f; }
 
 // =======================================================================================
 // Useful fake containers
